add error handling case, improve code structure
