+++ app/src/main/java/com/ece452/pillmaster/model/Pill.kt
@@ -8,6 +8,5 @@ data class Pill(
     var userId: String = "", // Carereceiver userId the Pill belongs to
     var name: String = "", // name of the Pill, generated from AddReminder
     var description: String = "", // description of Pill user enters, generated from AddReminder
-    var info: String = "TODO", // detailed info, compatibility info of the Pill, TODO: Retrieve from Cloud Function
-    var incompatibleDrugs: String? = "Unknown",
+    var info: String? = "Unknown", // detailed info, compatibility info of the Pill,
 )
